move tests to a separate directory
since I hope there will be lots of them, I move the test on the gtfs_rt
to a separate directory
