+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod test_utils;
+mod update_model_test;
